Added Logging (#88)
