+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "rbxlx-to-rojo"
-version = "1.0.1"
+version = "1.1.0"
 authors = ["Kampfkarren"]
 edition = "2024"
