+++ src/main.rs
@@ -1,100 +1,19 @@
 #![cfg_attr(not(debug_assertions), windows_subsystem = "windows")]
 
 use log::info;
-use rbxlx_to_rojo::{filesystem::FileSystem, process_instructions};
+use rbxlx_to_rojo::{
+    filesystem::FileSystem,
+    process_instructions,
+    utils::{Problem, setup_logger},
+};
 use std::{
     borrow::Cow,
-    env::{self},
-    fmt, fs,
-    io::{self, BufReader, Write},
-    sync::{Arc, RwLock},
+    fs,
+    io::BufReader,
 };
 
-#[derive(Debug)]
-enum Problem {
-    BinaryDecodeError(rbx_binary::DecodeError),
-    InvalidFile,
-    IoError(&'static str, io::Error),
-    FileDialogueCancel,
-    FileDialogueError(String),
-    XMLDecodeError(rbx_xml::DecodeError),
-}
-
-impl fmt::Display for Problem {
-    fn fmt(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
-        match self {
-            Problem::BinaryDecodeError(error) => write!(
-                formatter,
-                "While attempting to decode the place file, at {} rbx_binary didn't know what to do",
-                error,
-            ),
-
-            Problem::InvalidFile => {
-                write!(
-                    formatter,
-                    "The file provided does not have a recognized file extension"
-                )
-            }
-
-            Problem::IoError(doing_what, error) => {
-                write!(formatter, "While attempting to {}, {}", doing_what, error)
-            }
-            Problem::FileDialogueCancel => write!(formatter, "Didn't choose a file."),
-
-            Problem::FileDialogueError(error) => write!(
-                formatter,
-                "Something went wrong when choosing a file: {}",
-                error,
-            ),
-
-            Problem::XMLDecodeError(error) => write!(
-                formatter,
-                "While attempting to decode the place file, at {} rbx_xml didn't know what to do",
-                error,
-            ),
-        }
-    }
-}
-
-struct WrappedLogger {
-    log: env_logger::Logger,
-    log_file: Arc<RwLock<Option<fs::File>>>,
-}
-
-impl log::Log for WrappedLogger {
-    fn enabled(&self, metadata: &log::Metadata) -> bool {
-        self.log.enabled(metadata)
-    }
-
-    fn log(&self, record: &log::Record) {
-        if self.enabled(record.metadata()) {
-            self.log.log(record);
-
-            if let Some(log_file) = &mut *self.log_file.write().unwrap() {
-                log_file
-                    .write_all(format!("{}\r\n", record.args()).as_bytes())
-                    .ok();
-            }
-        }
-    }
-
-    fn flush(&self) {}
-}
-
 fn routine() -> Result<(), Problem> {
-    let env_logger = env_logger::Builder::new()
-        .filter_level(log::LevelFilter::Info)
-        .build();
-
-    let log_file = Arc::new(RwLock::new(None));
-    let logger = WrappedLogger {
-        log: env_logger,
-        log_file: Arc::clone(&log_file),
-    };
-
-    log::set_boxed_logger(Box::new(logger)).unwrap();
-    log::set_max_level(log::LevelFilter::Info);
-
+    let log_file = setup_logger();
     info!("rbxlx-to-rojo {}", env!("CARGO_PKG_VERSION"));
 
     info!("Select a place file.");
@@ -161,15 +80,15 @@ fn routine() -> Result<(), Problem> {
             .map_err(|error| Problem::IoError("couldn't create log file", error))?,
     );
 
-    info!("Starting processing, please wait a bit...");
     process_instructions(&tree, &mut filesystem);
-    info!("Done! Check rbxlx-to-rojo.log for a full log.");
     Ok(())
 }
 
 fn main() {
     #[cfg(feature = "gui")]
     {
+        use std::env;
+
         let args = env::args().nth(1);
         if args.is_none() {
             if let Err(error) = rbxlx_to_rojo::gui::run() {
